Color icons, not buttons
when an icon color is set, it tried to color the button and the icon
causing often ureadable results. now color just the icon but not
the button, giving a less magic behavior. it won't be possible
to color non monochrome icons, but that's kinda expected and should
be an acceptable compromise

BUG:429399
BUG:429972