Don't hide list item separator on hover when background is transparent
A couple of list items set activeBackgroundColor: to "transparent"
because having a visual hover effect doesn't make sense but they don't
want to set hoverEnabled: to false due to content items that require
hover effects.

However in this case, the list item's separator disappears when the
cursor passes over it, which only makes sense when the hover effect is
visible. So let's handle this case as well.