[Units] Reduce shortDuration and longDuration by 50ms
They were a bit too long and made animations feel slightly sluggish in many of the situations that they were used in.

It is not currently possible to force each duration to last a consistent amount of frames.
That probably wouldn't be desireable anyway since user inputs are not perfectly synced to the display's refresh rate.

- veryShortDuration has the duration of 3 frames at 60Hz
- shortDuration has the duration of 6 frames at 60Hz
- longDuration has the duration of 12 frames at 60Hz