Don't manually call reverseTwinsChanged when destructing FormLayout
This line causes a crash when destructing the screenlocker KCM.
It doesn't seem to be caused by reverseTwinsChanged itself but rather by
accessing the Qt object. I don't fully undestand the root cause.

However the call itself appears to be unnecessary since in the line
above we write to reverseTwins and that triggers the changed signal if
needed.

BUG: 431419
BUG: 428461