Replace PlatformThemePrivate with an implementation that shares data
In an average application, there will be a lot of PlatformTheme
instances, since they will be attached to almost every object. However,
PlatformThemePrivate is currently almost 800 bytes large. Along with
some other things, this means we are using a lot of memory for each
PlatformTheme instance.

This changes things by having the data shared between instances of
PlatformTheme when they inherit. This greatly reduces the number of
actual data objects that are needed, since the majority of PlatformTheme
instances will inherit.