Completely rework BasicTheme
We now have a C++ side object definition that should be used as base for
custom Theme objects. This greatly simplifies the C++ side logic and
makes it very explicit what can and cannot be defined. In addition, we
look up the Theme file directly rather than having some inline QML code
that relies on a feature that is going away (minor versioning) to work.