Do not load Theme.qml as Theme singleton, instead use BasicThemeDefinition
The Theme singleton has been deprecated for a long time. To not
completely break things that may still be using it, this will return an
instance of BasicThemeDefinition as Theme singleton, which should at
least give some colors.

Unfortunately, doing this means older Theme.qml files fail to load in
BasicTheme as QQmlComponent will refuse to load anything with `pragma
Singleton` in it. To workaround this, we strip the file of any such
lines before loading it.