Change and improve PlaceholderMessage new API
The way this was done involved an unnecessary compatibility item. We can
implement the feature differently, in a way that avoids this to save
some memory.

Also re-use the same visual styling (lighter text, large title) as the
existing implementation, instead of the new visually incompatible
worse-looking style that was added in the previous API change.

Also change one of the API documentation examples to use the new API.