Split PageTab out of SwipeNavigator
This style of tab is useful outside of the SwipeNavigator;
some applications have taken to reeimplenting it because
they don't want to use the full SwipeNavigator. Splitting
it out into public API makes it so that these applications
can use a consistent tab style.

Code was also cleaned up somewhat.