Add a note about humanMoment not being for animation durations
This adds in an explicit note to say that the humanMoment unit should not be used
to determine the duration of an animation, but rather it should be used
to determine when something has taken long enough that the user should
expect some kind of feedback. This came up on another MR(1), and it
seemed prudent to add an explicit note (rather than trusting the
implications of the brief to be sufficient discouragement).

1: https://invent.kde.org/frameworks/kirigami/-/merge_requests/270#note_217555