[BasicListItem] Don't unconditionally apply leading and trailing padding
Right now the leadingPadding and trailingPadding properties are being
unconditionally applied even if there are no leading or trailing items.
Let's only apply them if there are leading or trailing items. This fixes
the problem of BasicListItems witout leading items having their icons
indented too far to the right.

BUG: 435784
FIXED-IN: 5.82