Add KDE Store avatar and profile links to AboutPage
With the introduction of the direct-link avatar endpoint on the KDE Store, we are finally able to make use of the ocsUsername information. We were able to also do this for profile links in the past, but with the avatar endpoint there, we can now do both of them.

This adds a checkbox in case avatars exist, so that we don't go out and fetch things without the user's permission. Privacy is important, so we want to make sure to catch that.

Also further fixes a couple of warnings when using the data structure described in the AboutPage's documentation (rather than an actual KAboutData instance)