[BasicListItem] Offer vertical fill options for leading/trailing items
These items were previously always resizing the provided components to
vertically fill all available space. This was an undocumented behavior,
but some apps were relying on it.

I changed this in feb9fef03b42f56507282b59600c183963377133 to allow
custom leading trailing items that do not fill height, but in the
process broke the leading/trailing items in the Users KCM and Tok, which
were relying on this behavior, and that commit was reverted.

This commit introduces new properties to allow the developer to choose
whether leading or trailing items should fill the space vertically, or
use a developer-provided height so they can be a fixed size. For
backwards compatibility, it defaults to the old fill height behavior.

A KF6 TODO to remove them is added, because ideally you could just
specify a manual height to disable vertical filling, but that turns out
to not be easily implementable with the current code without breaking
users of it.