Fix non-atomic updating of PlatformTheme colours
When notifying PlatformTheme objects of changes, there is no need to
queue the event.

Queueing events is a bad design when it comes to propogating changes as
we can get frames rendered in the middle given us a broken state.

There is still an issue when it comes to propagating to children or if
colours are individually set, but this fixes at least one of the issues.