Only emit sync signal if we have finished construction
We cannot control what happens in slots connected to sync().
In particular our own Material style accesses the Theme attached
property in onSync. If we have not finished construction the engine
did not set the instance as the attachedObject and constructs a new
one. This eventually triggers syncToQml() again and so on, causing
a stack overflow.
BUG:438682