Add settings components
This commit adds three components: CategorizedSettings,
GenericSettingsPage and SettingAction.

CategorizedSettings provide a PageRow with a list, list items will
appear on the right of the list.

GenericSettingsPage is a thin abstraction that make it easy to create
simple settings pages.

SettingAction is what tells the CategorizedSettings what pages are
going to appear on the list.