Avoid creating then immediately discarding a page's ActionButton instance
During page creation, the globalToolBarStyle property may not yet be
properly initialised, which means we load and create ActionButton, then
almost immediately afterwards discard it again once we know for sure
we're using the ToolBar style.

This also unfolds the binding for "actionButtons.active" since it was
really hard to follow.