globaltoolbar: Use strict === equality in JavaScript
It is used for enumerations (int), counters (also int) and objects. So,
it should be safe to change them.

The only case where it is used for strings here, is Qt.platform.os which
is confirmed to be of a primitive string type.

(cherry-picked from commit 6970bcc8a36eddaab23c682be9e20cc06803602c)
(doing it right this time)

Also added comment on showNavigationButtons type and semantics, because
it is a mess and has been causing two breakages in past.