change color when card is highlighted
I didn't notice the highlighted part when changing DefaultCardBackground. Thanks to dimkard (from kirigami matrix) it was "brought to light" and so I made this MR.

Something i noticed in discover is that when directly going into Applications section it already highlights the first application as seen below. The reason is why i am bringing this up is because I don't know if this is a problem I made or it's just kirigami's implementation. 
![image](/uploads/a1a11a7f6840580fd4245f7b3e516eaf/image.png)