Use relative includes for own export macro header
The export header is in the same folder in the same install dir.
So relying on the POSIX standard to first check the same dir with ""
includes speeds up the header find, and also makes sure that
the header is found when the include path does not include the subdir
with the export header and the class headers, but only to the parentdir
matching the C++ namespace of the classes.

NO_CHANGELOG