Remove obsolete files and everything related to the QMake build system
With the changes made to auto-generate qrc/qml files through
ECMQmlModule, we no longer have to manually maintain those files. This
means they will be going out of sync quickly and the QMake build system
is going to regress even further. So rather than sort-of-supporting it,
just drop it completely. It's been de-facto deprecated for a while
anyway.

This also removes all the qmake related files.