Fix loading custom displayComponent inside ActionToolBar sub actions
The issue is that due to the Loader, it was not possible to close
automatically the menu once clicking on an menu item.

This is an API break because the previous api accepted everything but didn't
really work. Now only elements who have a action property are accepted
but in really the menu will only work correctly with QQC2.MenuItem.