Units: add mediumSpacing property
mediumSpacing is `std::round(smallSpacing * 1.5)`. It is meant to be
used in toolbars or toolbar-like layouts and is intended to improve
visual consistency with our Qt Widget apps. With Noto Sans 10pts (the
current default font), this is equivalent to 6px, which is the same as
the default amount of spacing and margins for layouts in Qt Widgets with
the Breeze QStyle.