Temporarily disable shader code to unblock the Qt6 build
For now this will use the already existing software fallback for the
shadow elements. It is supposed to fail if we forget to remove this
again in time for an actual KF6 release.

This is ugly, but was deemed acceptable during the last KF6 meeting in
order to unblock all other modules that need a building Kirigami. This
will of course not actually make Kirigami work, the QML changes needed
for that are still missing.