FormLayout: make section headings slightly smaller but bolder
Using size alone does not really succeed in making section headings
stand out from the groups they are the header for, resulting in a
visually muddy appearance. This is exactly the use case envisioned for
the Kirigami.Heading.Type.Primary Heading type and is explicitly called
out in the documentation, so let's use it here to make section headings
pop a bit more.

No longer trying to rely on size alone lets us reduce headings' font
size a bit, which makes the size hierarchy more typographically
correct, with level-5-sized text headed by a level 3 heading, not a
level 2 heading.