Render Page's default heading as PlainText
Fixes problems when Text.AutoText thinks the text should be rendered as
html when it really shouldn't. This does remove the possibility of using
html intentionally, however that doesn't work well anyways and I don't
really see a usecase for it. If some app really needs that behavior, it
can use it with a custom header delegate