InlineMessage: Do not layout for actions at bottom when there are none
A lot of the layouting depends on the `actions.atBottom` property
which is set when they cannot be accomodated in a single line, so
the action toolbar is placed below the text.

This fixes the case when the text occupies several lines, but there
are no actions, preventing an extra bottom margin and keeping the
items centered.