BannerImage: Delay recalculating title size until after anchors have settled
The order of binding evaluation is not determined. This means that we
can sometimes end up in a state where both left and right anchors are
set. In that case, the item ends up being resized to full width first
and then the other anchor is released, but the size is still wrong. This
then means alignment is broken. To fix this, we need to tell QtQuick to
recalculate the size once the anchors have settled.