Require passing tests on BSD
There have been several times mr have been merged breaking tests,
sometimes near to release. require passing tests on any mrin order to
be able to merge.
Uses only BSD for now as not all things are correctly ported to Qt6 yet there is still things that are expected to fail on qt6 which can't be excluded for the Linux target