SearchField: Align search icon vertically
In case if for some reason a text field is taller than its normal
implicit height, let the icon be centered instead of sticking it to the
bottom. The verticalCenterOffset binding takes care of vertical padding,
and visually synchronizes the icon with the text placeholder for any
integer value of top and bottom padding.

Also get rid of extra bindings to parent change by referring directly to
root component by its id. We know searchIcon is designed to always stay
there, never to be reparented.