SwipeListItem: Make sure we only show the aggressive warning when it's due
I've seen at times that the warning would show in Discover despite the
conditions explained in the text would not be matched. This is because
of how Repeater works and ultimately the size here.

Instead of keeping it in a property that needs to keep being updated, do
it in a local function so at least we are calculating it on the spot
rather than using the last update of the property.

BUG: 455789