Don't load page header components asynchronously by default
This usually isn't noticed on desktop because computers are generally too fast for it to be noticeable, but on slower hardware (phones), loading the header components asynchronously looks unsatisfactory.

Headers generally consist of a text title, and the action toolbar. The action toolbar already does async loading, so we are basically just loading the title here async. I think that it should be loaded synchronously along with the page, and if there are custom components in the header, it should be up to the developer to load it async themselves.

Separately I think work does need to be done with the action toolbar layout to make it look less jumpy as it is loading, but I think this is a good start to prevent the jumpy title load in on mobile.