InlineMessage: Don't bind your logic to the value of .visible property
Lesson learnt from the experience with Battery & Brightness inhibition
hints list. Although, there it was a visibleChildren property which
messed us up by turning into an empty list when a whole component is
not visible. Anyway, same here: don't bind sizing calculations to the
value of closeButton.visible, as it won't be the same as the value of
what it is in turn bound to -- because when a whole InlineMessage is
hidden, the value `visible: false` will implicitly propagate down the
tree, but would be restored to an explicitly bound one when shown again.

It does not fix anything noticeable at this moment, but that's because
half of animations are not working at all atm: hiding is instant.