FormLayout: remove excessive spacing above unlabeled separators
The clear intention of the code (as even evidenced by a code comment)
was to add more spacing above the label of a labeled separator item
being used as a section header. But the code was not checking for
whether the separator actually had a label, so the extra spacing log
was being inappropriately applied to all separator items.

BUG: 462300
FIXED-IN: 5.102