NavigationTabBar: Move button width calculation out of NavigationTabButton
It is really strange to set the button widths from the button, since it makes the assumption that the parent is a NavigationTabBar.

Move this width calculation to the tab bar instead.

I'm not sure if this is something that needs to be done in KF6 due to defaults changing and an added property, but there aren't any uses of NavigationTabButton directly (it's always instantiated from within the NavigationTabBar).