InlineMessage: Downstream and clarify padding expressions that are based on background's border width
Template implementation assumed too much that background item must be a
Rectangle (is it pretty much the only base type that has border.width
property). Not only that, but relying on border.width didn't make any
sense, since border is drawn by other means (multiple Rectangles with
offsets) and it just happened to match the default width value `1` of
an invisible border.