ActionMenuItem: Shorten one binding expression, remove useless check
Checking for existence of an attached property in QML is pointless due
to the fact that engine would always construct an attached object the
first time it's accessed.

Also, it's quite strange that any of this works with 2.3 version of QQC2
import, as documentation claims that ActionGroup is `Since: Qt 5.10`.
But I checked it on a separate isolated file, and even normal & attached
forms of object declaration worked fine.