Clean up examples
This includes adding/fixing import aliases, bumping import versions,
fixing code style, replacing multiline "strings" with modern `template
literals` and more.

Some examples aren't usable as-is, some are not clear (to me) as to what
they are supposed to showcase. But figuring that out is a topic for
another time.