ListSectionHeader: Refactor layout to fix elision, add tooltip, add tests
Tried to make it work with collapsible Kirigami.ActionToolBar, but seems
like there's no way to make it work that way with current layouts API.

Added ToolTip to the delegate as a whole. Its behavior is suboptimal
when there is custom content with its own tooltip, but adding an enabled
HoverHandler is not an option due to QTBUG-106489.

BUG: 464681
FIXED-IN: 5.103