ToolBarLayout: Split up implicit size calculation and layouting
If an explicit height is set, we need to make sure to use it for
delegate's height. Unfortunately, Control unconditionally sets the
contentItem's height, which makes QQuickItem::heightValid() rather
useless. By separating the layouting from the size calculation, we can
have the layouting make use of the width/height of the item rather than
relying on previously calculated values, making it always use the right
value.

BUG: 464910