Stop MouseArea from eating the first hover input event
We use HoverHandler to detect whether a mouse was used, however since
hoverEnabled was set on the MouseArea, it was eating the first event and
enabling users to drag the dialog by accident.

Also, by default acceptedButtons is LeftButton, so let's enable the
MouseArea for RightButton, so you can't accidentally drag dialogs by
right- clicking, which is common with text inputs.

BUG: 465394