Set fallback theme path when a custom icon theme is used
When a custom icon theme is used, built-in Breeze icons are unavailable.
This is due to some undocumented Qt behavior regarding loosely organized
icon files, so when the theme name is explicitly set, make sure to add the
icons directory to the fallback search paths so the icons still work.