Default page categorized settings
Add the ability to change the default page that categorized settings opens to. Designed to be used to allow fine grained options that open the settings on a set page rather than generically opening at the top one all the time.

Example that make use of this in neochat network/neochat!776