fix deadlock when talking to portal service
when the portal service itself uses kirigami's tablet mode watcher (as
is the case currently through the file open dialog -> breeze ->
kirigami) then we end up in a deadlock situation since we are talking to
the xdg-desktop-portal and that talks to xdg-desktop-portal-kde (i.e.
the process that issued the call). Qt has magic to prevent a process
deadlocking when talking to itself, so in the edge case scenario that
the apppid==dbusservicepid we'll instead talk to the impl.portal
interface directly (i.e. itself) and bypass xdg-desktop-portal entirely.

since impl.portal is technically a different interface we are generating
a second interface class for this purpose and template-use either the
regular interface or the impl interface. they are API compatible right
now.

despite this currently only being caused by breeze the fix is
implemented globally in kirigami because technically nothing prevents
the qml-only file open dialogs in the portal from using kirigami
directly either.