Draft: Remove checkable, checked and enabled properties from FormLayout/FormData
This removes all three properties from FormData attached type and their
usages from FormLayout component. The reason is that they are not used,
require considerable amount of code to keep it from falling apart, and
requires FormLayout override/style CheckBox component which is really
not its job and which we miserably failed to maintain consistent at
least with qqc2-desktop-style (and which also is unreadable in RTL mode).

The checkable/checked properties are not used anywhere throughout Plasma
software. The Kirigami.FormData.enabled only appears twice in
plasma-firewall for CheckBox labels (where it doesn't do anything,
because according to my observations disabled labels don't change colors
in qqc2-desktop-style).