SelectableLabel: Prevent it from participating in Tab chain
This is a piece of non-editable text, it doesn't need to be focused via
keyboard, especially since it doesn't provide any visual feedback. Even
when it contains clickable hyperlinks, unfortunately they are not
usable from keyboard. Basically the only thing you can do with a
readOnly TextArea via keyboard is Ctrl+A Ctrl+C.