Avoid initial Search animation
The search bar has a feature where it shows an icon if it does
not have activeFocus.

This runs on startup as QML is evaluated before the window is
visible. When the window is finally shown, the focus is set correctly,
but at that point it's considered a change and the animations play.

This means the placeholder text can be seen moving on the first open
which looks glitchy.