Drop aged TODO comment
It was referring to the fact that signal handlers with parameters may
be directly connected to a setter slot with matching parameter type.
Apparently it was never used this way: the only two lines of code that
refer to this signal in C++ code, they do not use actual colors value,
they only use the fact that the signal itself is triggered.