ActionTextField: Disable shortcut for invisible and disabled text fields
It doesn't make sense to intercept shortcuts when it can't be
meaningfully handled. In fact, it interferes and breaks multiple
form-factors scenario where one search field might be in a top bar and
another one in a sidebar and only one is visible or usable at once,
such as Kirigami Gallery main view and a global drawer.