Avatar: Port `actions` group property members to QObject* and manage lifecycle
It is important to emit change signals whenever a QObject member is
destroyed. MEMBER syntax of Q_PROPERTY won't manage that for you.

Also, main and secondary actions were meant to be Kirigami.Actions, but
since Kirigami.Action is a pure QML type and QQC2.Action is in a
private header, QObject* is unfortunately as close as we can get to its
real type. Which is still better than QVariant, and gives us some
stronger nullability guarantees (e.g. can't be `undefined` or other
meaningless type).