avatar: avoid shadowing property smooth
Smooth is a property of QQuickItem. In 6.6 these properties are declared
final which causes an error when used.

There's also no benefit to not setting it on an image. There's
effectively no performance gain, and no use-case for an image being
pixelated. For that reason setting the property now does nothing.