Port i18n to qsTr
We can't depend on KI18n, so we can't assume i18n() is present.
Instead, use qsTr, which comes with Qt and is what we use for
translation in Tier-1 frameworks.

These are the only i18n calls remaining. The rest the search results are
either in doc comments or in examples, which is acceptable.