PageRow: Remove leftSidebar property
The leftSidebar property is not used anywhere in any project. The only
mention outside PageRow component is anchors margins binding in
AbstractApplicationWindow. Moreover, sidebarControl component in the
column view layout references nonexistent property `root.leftDrawer`,
which means it's been broken for a while now with no one complaining;
and it's been like that since the introduction of leftSidebar in
502a7ed20c2aae29449df71a3b1f82c750e0748c.